+++ R/create_report.R
@@ -14,7 +14,8 @@
 #' @family utils
 #' @export
 #' @importFrom assertthat assert_that is.string noNA
-#' @importFrom checklist ask_yes_no menu_first read_checklist use_author
+#' @importFrom checklist ask_yes_no menu_first organisation read_checklist
+#' use_author
 #' @importFrom fs dir_create file_copy is_dir is_file path
 #' @importFrom gert git_find
 create_report <- function(path = ".", shortname) {
@@ -37,6 +38,16 @@ msg = "The report name folder may only contain lower case letters, digits and _"
     msg = "The report name folder already exists."
   )
 
+  lang <- c(nl = "Dutch", en = "English", fr = "French")
+  lang <- names(lang)[
+    menu_first(lang, title = "What is the main language of the report?")
+  ]
+  style <- c("INBO", "Vlaanderen")
+  style <- ifelse(
+    lang != "nl", "Flanders",
+    style[menu_first(style, title = "Which style to use?")]
+  )
+
   # build new yaml
   readline(prompt = "Enter the title: ") |>
     gsub(pattern = "[\"|']", replacement = "") |>
@@ -47,10 +58,10 @@ msg = "The report name folder may only contain lower case letters, digits and _"
     gsub(pattern = "[\"|']", replacement = "") -> subtitle
   yaml <- c(yaml, paste("subtitle:", subtitle)[subtitle != ""])
   cat("Please select the corresponding author")
-  authors <- use_author()
+  authors <- check_author(lang = lang)
   c(yaml, "author:", author2yaml(authors, corresponding = TRUE)) -> yaml
   while (isTRUE(ask_yes_no("Add another author?", default = FALSE))) {
-    author <- use_author()
+    author <- check_author(lang = lang)
     authors[, c("given", "family", "email")] |>
       rbind(author[, c("given", "family", "email")]) |>
       anyDuplicated() -> duplo
@@ -66,7 +77,7 @@ msg = "The report name folder may only contain lower case letters, digits and _"
   cat("Please select the reviewer")
   duplo <- 1
   while (duplo > 0) {
-    author <- use_author()
+    author <- check_author(lang = lang)
     authors[, c("given", "family", "email")] |>
       rbind(author[, c("given", "family", "email")]) |>
       anyDuplicated() -> duplo
@@ -80,6 +91,8 @@ msg = "The report name folder may only contain lower case letters, digits and _"
   lang <- names(lang)[
     menu_first(lang, title = "What is the main language of the report?")
   ]
+  org <- organisation$new()
+  aff <- org$get_organisation[["inbo.be"]]$affiliation[lang]
   style <- c("INBO", "Vlaanderen")
   style <- ifelse(
     lang != "nl", "Flanders",
@@ -106,8 +119,7 @@ msg = "The report name folder may only contain lower case letters, digits and _"
     "lof: TRUE"[ask_yes_no("Do you want a list of figures?", default = FALSE)],
     "lot: TRUE"[ask_yes_no("Do you want a list of tables?", default = FALSE)],
     keywords, "community: \"inbo\"", "publication_type: report",
-    "funder: Research Institute for Nature and Forest (INBO)",
-    "rightsholder: Research Institute for Nature and Forest (INBO)",
+    paste("funder:", aff), paste("rightsholder:", aff),
     "bibliography: references.bib", "link-citations: TRUE",
     "site: bookdown::bookdown_site", "output:", "  INBOmd::gitbook: default",
     "  INBOmd::pdf_report: default", "  INBOmd::epub_book: default",
@@ -245,7 +257,7 @@ author2yaml <- function(author, corresponding = FALSE) {
     yaml <- c(yaml, sprintf("    orcid: \"%s\"", author$orcid))
   }
   if (!is.na(author$affiliation) && author$affiliation != "") {
-    yaml <- c(yaml, paste("    affiliation: \"%s\"", author$affiliation))
+    yaml <- c(yaml, sprintf("    affiliation: \"%s\"", author$affiliation))
   }
   if (!corresponding) {
     return(paste(yaml, collapse = "\n"))
@@ -279,4 +291,23 @@ add_address <- function(type = "client") {
     sprintf("%s:", type), sprintf("  - %s", address),
     sprintf("%s_logo: %s", type, logo)[logo != ""]
   )
+}
+
+#' @importFrom checklist use_author
+check_author <- function(lang = "nl") {
+  person <- use_author()
+  if (is_inbo(person) && !person$affiliation %in% inbo_affiliation) {
+    paste0("`", inbo_affiliation, "`", collapse = "; ") |>
+      sprintf(fmt = "INBO staff must use one of %s as affiliation.") |>
+      cat("Please update the author information.", "\n", sep = " ")
+    person <- check_author(lang = "nl")
+    person$affiliation <- inbo_affiliation[lang]
+  }
+  if (is_inbo(person) && is.na(person$orcid)) {
+    cat(
+      "INBO staff must provide an ORCID.",
+      "Please update the author information.", "\n", sep = " "
+    )
+    person <- check_author(lang = "nl")
+  }
 }