+++ R/pdf_report.R
@@ -25,50 +25,70 @@ pdf_report <- function(
     yaml_front_matter() |>
     validate_persons(reviewer = TRUE) |>
     validate_rightsholder() -> fm
-  floatbarrier <- ifelse(has_name(fm, "floatbarrier"), fm$floatbarrier, NA)
-  assert_that(length(floatbarrier) == 1)
+  stopifnot(
+    "`internal` option in yaml is not allowed" = !has_name(fm, "internal"),
+    "`pagefootmessage` option in yaml is not allowed" =
+      !has_name(fm, "pagefootmessage")
+  )
+  style <- list(INBO = "nl", Vlaanderen = "nl", Flanders = c("en", "fr"))
+  languages <- c(nl = "dutch", en = "english", fr = "french")
+  hide_defaults <- list(internal = FALSE, lof = FALSE, lot = FALSE)
+  defaults <- c(
+    hide_defaults, style = names(style)[1], public_report = TRUE,
+# use the first language of the style when set
+# otherwise use the first language of the first style
+    lang = unlist(style[fm$style]) |>
+      c(style[[1]]) |>
+      head(1) |>
+      unname(),
+    floatbarrier = NA, watermark = NULL,
+    other_lang = list(names(languages))
+  )
+  extra <- !names(defaults) %in% names(fm)
+  fm <- c(fm, defaults[extra])
+
+  assert_that(length(fm$floatbarrier) == 1)
   assert_that(
-    floatbarrier %in% c(NA, "section", "subsection", "subsubsection"),
+    fm$floatbarrier %in% c(NA, "section", "subsection", "subsubsection"),
     msg =
 "Allowed options for `floatbarrier` are missing, 'section', 'subsection' and
 'subsubsection'"
   )
-  style <- ifelse(has_name(fm, "style"), fm$style, "INBO")
   stopifnot(
-    "`style` is not a string" = is.string(style),
-    "`style` is not a string" = noNA(style),
-    "`style` must be one of 'INBO', 'Vlaanderen' or 'Flanders'" =
-      style %in% c("INBO", "Vlaanderen", "Flanders")
+    "`style` is not a string" = is.string(fm$style),
+    "`style` is not a string" = noNA(fm$style),
+    "`style` must be one of 'INBO', 'Vlaanderen', 'Flanders'" =
+      fm$style %in% names(style)
   )
-  lang <- ifelse(
-    has_name(fm, "lang"), fm$lang, ifelse(style == "Flanders", "en", "nl")
-  )
-  assert_that(length(lang) == 1)
-  languages <- c(nl = "dutch", en = "english", fr = "french")
+  assert_that(length(fm$lang) == 1)
   assert_that(
-    lang %in% names(languages),
+    fm$lang %in% names(languages),
     msg = paste(
       "`lang` must be one of:",
       paste(sprintf("'%s' (%s)", names(languages), languages), collapse = ", ")
     )
   )
   assert_that(
-    style != "Flanders" || lang != "nl",
-    msg = "Use style: Vlaanderen when the main language is Dutch"
+    fm$lang %in% style[[fm$style]],
+    msg = vapply(
+      names(style), FUN.VALUE = character(1), style = style,
+      FUN = function(s, style) {
+        sprintf("`%s`", style[[s]]) |>
+          paste(collapse = ", ") |>
+          sprintf(fmt = "%2$s: %1$s", s)
+      }
+    ) |>
+      paste(collapse = "; ") |>
+      sprintf(fmt = "Available combinations of `style` and `lang`\n%s")
   )
-  other_lang <- fm$other_lang
-  if (is.null(other_lang)) {
-    other_lang <- names(languages)
-  }
-  other_lang <- other_lang[other_lang != lang]
+  fm$other_lang <- fm$other_lang[fm$other_lang != fm$lang]
   assert_that(
-    all(other_lang %in% names(languages)),
+    all(fm$other_lang %in% names(languages)),
     msg = paste(
       "all `other_lang` must be in this list:",
       paste(sprintf("'%s' (%s)", names(languages), languages), collapse = ", ")
     )
   )
-  validate_doi(ifelse(has_name(fm, "doi"), fm$doi, "1.1/1"))
 
   path("pandoc", "inbo_rapport.tex") |>
     system.file(package = "INBOmd") -> template
@@ -76,22 +96,8 @@ pdf_report <- function(
     "research-institute-for-nature-and-forest.csl", package = "INBOmd"
   )
 
-  style <- ifelse(style == "Flanders" & lang == "fr", "Flandre", style)
-
   args <- c(
-    "--template", template, pandoc_variable_arg("documentclass", "report"),
-    switch(
-      style,
-      Flanders = pandoc_variable_arg("style", "flanders_report"),
-      Flandre = pandoc_variable_arg("style", "flandre_report"),
-      Vlaanderen = pandoc_variable_arg("style", "vlaanderen_report"),
-      INBO = pandoc_variable_arg("style", "inbo_report")
-    ),
-    pandoc_variable_arg("corresponding", fm$corresponding),
-    pandoc_variable_arg("shortauthor", gsub("\\&", "\\\\&", fm$shortauthor)),
-    pandoc_variable_arg(
-      "babel", paste(languages[c(other_lang, lang)], collapse = ",")
-    ),
+    "--template", template,
     ifelse(
       compareVersion(as.character(pandoc_version()), "2") < 0,
       "--latex-engine", "--pdf-engine"
@@ -101,14 +107,60 @@ pdf_report <- function(
   )
   args <- args[args != ""]
 
-  if (has_name(fm, "lof") && isTRUE(fm$lof)) {
-    args <- c(args, pandoc_variable_arg("lof", TRUE))
+  draft <- !all(c("cover_description", "year") %in% names(fm))
+  if (!fm$public_report) {
+    Sys.time() |>
+      format("%Y-%m-%d %H:%M:%S") |>
+      c(fm$reportnr) |>
+      tail(1) -> fm$pagefootmessage
+    fm$internal <- TRUE
+  } else {
+    if (has_name(fm, "doi")) {
+      validate_doi(fm$doi)
+      draft <- draft && !all(c("depotnr", "doi", "reportnr") %in% names(fm))
+    } else {
+      draft <- TRUE
+      fm$doi <- "!!! missing DOI !!!"
+    }
+    fm$pagefootmessage <- fm$doi
   }
-  if (has_name(fm, "lot") && isTRUE(fm$lot)) {
-    args <- c(args, pandoc_variable_arg("lot", TRUE))
+  if (draft) {
+    c(en = "DRAFT", fr = "CONCEPTION", nl = "ONTWERP")[fm$lang] |>
+      c(fm$watermark) |>
+      paste(collapse = "\\\\") -> fm$watermark
   }
+
+  fm[names(fm) %in% names(hide_defaults)] |>
+    unlist() -> to_hide
+  fm[names(to_hide)[!to_hide]] <- NULL
+  var_arg <- c(
+    documentclass = "report",
+    style = c(
+      Flanders_en = "flanders_report", Flanders_fr = "flandre_report",
+      Vlaanderen_nl = "vlaanderen_report", INBO_nl = "inbo_report"
+    )[paste(fm$style, fm$lang, sep = "_")] |>
+      unname(),
+    fm[
+      c(
+        "corresponding", "doi", "internal", "lof", "lot", "watermark",
+        "pagefootmessage"
+      )
+    ],
+    shortauthor = gsub("\\&", "\\\\&", fm$shortauthor),
+    babel = paste(languages[c(fm$other_lang, fm$lang)], collapse = ",")
+  )
+  var_arg <- var_arg[!vapply(var_arg, is.null, logical(1))]
+  vapply(
+    seq_along(var_arg), FUN.VALUE = character(2), var_arg = var_arg,
+    FUN = function(i, var_arg) {
+      pandoc_variable_arg(names(var_arg)[i], var_arg[i])
+    }
+  ) |>
+    as.vector() |>
+    c(args) -> args
+
   vars <- switch(
-    floatbarrier, section = "", subsection = c("", "sub"),
+    fm$floatbarrier, section = "", subsection = c("", "sub"),
     subsubsection = c("", "sub", "subsub")
   )
   floating <- lapply(
@@ -250,7 +302,6 @@ validate_persons <- function(yaml, reviewer = TRUE) {
       c(tail(shortauthor, 1)) |>
       paste(collapse = " & ") -> yaml$shortauthor
   }
-  shortauthor <- paste(shortauthor, "<test.inbo.be>")
   assert_that(
     length(corresponding) == 1,
     msg = "A single corresponding author is required."
@@ -282,29 +333,54 @@ contact_person <- function(person) {
     gsub("\\1.", person$name$given, perl = TRUE) |>
     sprintf(fmt = "%2$s, %1$s", person$name$family) -> shortauthor
   if (!has_name(person, "orcid")) {
+    if (is_inbo(person)) {
+      sprintf(
+        "`orcid` required for %s %s", person$name$given, person$name$family
+      ) |>
+        stop(call. = FALSE)
+    }
     sprintf(
       "No `orcid` found for %s %s", person$name$given, person$name$family
     ) |>
       warning(call. = FALSE)
   }
   if (!has_name(person, "affiliation")) {
+    if (is_inbo(person)) {
+      sprintf(
+        "`affiliation` required for %s %s.\nMust be one of %s",
+        person$name$given, person$name$family,
+        paste0("`", inbo_affiliation, "`", collapse = "; ")
+      ) |>
+        stop(call. = FALSE)
+    }
     sprintf(
       "No `affiliation` found for %s %s", person$name$given, person$name$family
     ) |>
       warning(call. = FALSE)
+  } else {
+    if (is_inbo(person) && !person$affiliation %in% inbo_affiliation) {
+      sprintf(
+        "`affiliation` for %s %s must be one of %s", person$name$given,
+        person$name$family, paste0("`", inbo_affiliation, "`", collapse = "; ")
+      ) |>
+        stop(call. = FALSE)
+    }
   }
   if (is.null(person$corresponding) || !person$corresponding) {
     return(shortauthor)
   }
   assert_that(
-    "email" %in% names(person),
+    has_name(person, "email"),
     msg = "no `email` provided for the corresponding author."
   )
   sprintf("%s<%s>", shortauthor, person$email)
 }
 
 #' @importFrom assertthat assert_that has_name is.string noNA
+#' @importFrom checklist organisation
 validate_rightsholder <- function(yaml) {
+  org <- organisation$new()
+  aff <- org$get_organisation[["inbo.be"]]$affiliation
   stopifnot(
     "no `funder` found" = has_name(yaml, "funder"),
     "`funder` is not a string" = is.string(yaml$funder),
@@ -312,8 +388,6 @@ validate_rightsholder <- function(yaml) {
     "no `rightsholder` found" = has_name(yaml, "rightsholder"),
     "`rightsholder` is not a string" = is.string(yaml$rightsholder),
     "`rightsholder` is not a string" = noNA(yaml$rightsholder),
-"Research Institute for Nature and Forest (INBO) not defined as rightsholder" =
-      yaml$rightsholder == "Research Institute for Nature and Forest (INBO)",
     "no `community` found" = has_name(yaml, "community"),
     "`community` must be a string separated by `; `" =
       is.string(yaml$community),
@@ -321,6 +395,13 @@ validate_rightsholder <- function(yaml) {
       "inbo" %in% strsplit(yaml$community, "; "),
     "no `keywords` found" = has_name(yaml, "keywords"),
     "`keywords` must be a string separated by `; `" = is.string(yaml$keywords)
+  )
+  assert_that(
+    yaml$rightsholder %in% aff,
+    msg = sprintf(
+      "rightsholder must be one of the following\n%s",
+      paste(aff, collapse = "\n")
+    )
   )
   return(yaml)
 }
@@ -344,4 +425,12 @@ check_license <- function() {
       file_copy(license_file, overwrite = TRUE)
   }
   return(invisible(NULL))
+}
+
+#' @importFrom assertthat has_name
+is_inbo <- function(person) {
+  if (!has_name(person, "email")) {
+    return(FALSE)
+  }
+  grepl("inbo.be$", person$email, ignore.case = TRUE)
 }